+++ src/clients/__tests__/ClientList.test.tsx
@@ -2,9 +2,9 @@ import React from "react";
 import { render } from "@testing-library/react";
 
 import clientMock from "./mock-clients.json";
-import { ClientList } from "./ClientList";
 import { I18nextProvider } from "react-i18next";
-import { i18n } from "../i18n";
+import i18n from "../../i18n";
+import { ClientList } from "../ClientList";
 
 test("renders ClientList", () => {
   const { getByText } = render(
