+++ src/clients/add/__tests__/mock-serverinfo.json
