+++ src/components/external-link/__tests__/ExternalLink.test.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { render } from "@testing-library/react";
-import { ExternalLink } from "./ExternalLink";
+import { ExternalLink } from "../ExternalLink";
 
 describe("<ExternalLink />", () => {
   it("render with link", () => {
