+++ src/components/help-enabler/HelpHeader.tsx
@@ -6,8 +6,6 @@ import {
   Split,
   SplitItem,
   Switch,
-  Text,
-  TextVariants,
 } from "@patternfly/react-core";
 import { Trans, useTranslation } from "react-i18next";
 import { HelpIcon, ExternalLinkAltIcon } from "@patternfly/react-icons";
