+++ src/components/help-enabler/__tests__/HelpItem.test.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { render } from "@testing-library/react";
 
-import { HelpItem } from "./HelpItem";
+import { HelpItem } from "../HelpItem";
 
 describe("<HelpItem />", () => {
   it("render", () => {
