+++ src/components/realm-selector/__tests__/RealmSelector.test.tsx
@@ -2,10 +2,7 @@ import React from "react";
 import { mount } from "enzyme";
 import { act } from "@testing-library/react";
 
-import { AlertPanel } from "../alert/AlertPanel";
-import { useAlerts } from "../alert/Alerts";
-
-import { RealmSelector } from "./RealmSelector";
+import { RealmSelector } from "../RealmSelector";
 
 it("renders realm selector", async () => {
   const wrapper = mount(
