+++ src/help.json
@@ -1,7 +1,5 @@
 {
-  "en": {
-    "help": {
-      "storybook": "Sometimes you need some help and it's nice when the app does that"
-    }
+  "help": {
+    "storybook": "Sometimes you need some help and it's nice when the app does that"
   }
 }
