+++ src/messages.json
@@ -1,24 +1,22 @@
 {
-  "en": {
-    "messages": {
-      "personalInfoHtmlTitle": "Personal Info",
-      "personalInfoIntroMessage": "Manage your basic information",
-      "Account Security": "Account Security",
-      "accountSecurityIntroMessage": "Control your password and account access",
-      "signingIn": "Signing In",
-      "device-activity": "Device Activity",
-      "applications": "Applications",
-      "applicationsIntroMessage": "Track and manage your app permission to access your account",
-      "fullName": "{{givenName}} {{familyName}}",
-      "unknownUser": "Anonymous",
-      "Keycloak Administration Console": "RH-SSO Administration Console",
-      "confirmImportClear": "If you clear this file, you need to click Browse button to re-import a valid file",
+  "messages": {
+    "personalInfoHtmlTitle": "Personal Info",
+    "personalInfoIntroMessage": "Manage your basic information",
+    "Account Security": "Account Security",
+    "accountSecurityIntroMessage": "Control your password and account access",
+    "signingIn": "Signing In",
+    "device-activity": "Device Activity",
+    "applications": "Applications",
+    "applicationsIntroMessage": "Track and manage your app permission to access your account",
+    "fullName": "{{givenName}} {{familyName}}",
+    "unknownUser": "Anonymous",
+    "Keycloak Administration Console": "RH-SSO Administration Console",
+    "confirmImportClear": "If you clear this file, you need to click Browse button to re-import a valid file",
 
-      "********* MASTHEAD *********": "",
-      "Sign out": "Sign out",
-      "Manage account": "Manage account",
-      "Server info": "Server info",
-      "Help": "Help"
-    }
+    "********* MASTHEAD *********": "",
+    "Sign out": "Sign out 2",
+    "Manage account": "Manage account 2",
+    "Server info": "Server info 2",
+    "Help": "Help!"
   }
 }
