+++ src/stories/HelpSystem.stories.tsx
@@ -14,7 +14,7 @@ import {
   HelpHeader,
 } from "../components/help-enabler/HelpHeader";
 import { I18nextProvider } from "react-i18next";
-import { i18n } from "../i18n";
+import i18n from "../i18n";
 
 export default {
   title: "Help System Example",
